Refactor modules select
